Add video endpoint to proxy url
Also do some general refactoring
