+++ src/constants.ts
@@ -1,0 +1,15 @@
+export const REDDIT_BASE_URL = 'https://www.reddit.com';
+export const CUSTOM_DOMAIN = 'rxddit.com';
+export const GITHUB_LINK = 'https://github.com/MinnDevelopment/fxreddit';
+export const USER_AGENT = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/116.0';
+
+export const FETCH_HEADERS = {
+    'User-Agent': USER_AGENT,
+    'Accept': 'application/json',
+    'Accept-Language': 'en-US,en;q=0.5',
+};
+
+export const RESPONSE_HEADERS = {
+    'Content-Type': 'text/html; charset=UTF-8',
+    'Cache-Control': 'public, max-age=86400'
+};
