+++ src/reddit/compile.ts
@@ -5,6 +5,7 @@ import { youtubeEmbed } from '../embeds/youtube';
 import { twitchClipEmbed } from '../embeds/twitch';
 import { twitterLinkEmbed } from '../embeds/twitter';
 import '../html';
+import { get_packaged_video } from '../util';
 
 function getDomainHandler(domain?: string) {
     switch (domain) {
@@ -54,29 +55,21 @@ export async function postToHtml(post: RedditPost): Promise<HTMLElement> {
             head.image(post.url, post.resolution?.width, post.resolution?.height);
             break;
         // case 'rich:video':
-        case 'hosted:video':
+        case 'hosted:video': {
             type = 'video.other';
 
-            try {
-                const url = new URL(`https://www.reddit.com${post.permalink}`);
-                const html = await fetch(url, {
-                    headers: {
-                        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/116.0',
-                    }, ...CACHE_CONFIG, redirect: 'follow' })
-                    .then(r => r.text()).then(parseHTML);
-                const player = html.querySelector('[packaged-media-json]');
-                const json = JSON.parse(player.getAttribute('packaged-media-json'));
-                const videos = json.playbackMp4s.permutations;
-
-                if (videos) {
-                    const lastVideo = videos[videos.length - 1];
-                    head.video(lastVideo.source.url, lastVideo.source.dimensions.width, lastVideo.source.dimensions.height);
-                }
-            } catch (ignored) {
+            const packagedVideo = await get_packaged_video(post.permalink);
+            if (packagedVideo) {
+                const { dimensions: { width, height } } = packagedVideo;
+                // head.video(packagedVideo.source.url, width, height);
+                // Proxied endpoint which resolves to the current video url, to avoid using expiring links
+                head.video(`/v${post.permalink}`, width, height);
+            } else {
                 // If we can't find a video with audio, we'll just settle with the one provided by Reddit
                 head.video(post.video_url ?? post.url, post.resolution?.width, post.resolution?.height);
             }
             break;
+        }
         case 'link':
         default: {
             const domainHandler = getDomainHandler(post.domain);
