+++ src/response_error.ts
@@ -1,0 +1,7 @@
+
+
+export default class ResponseError extends Error {
+    constructor(public status: number, public error: string) {
+        super(`${status}: ${error}`);
+    }
+}
